[#46] Presenter/Container 컴포넌트 구조 제거 (EventDetailPage)
